arboard

Struct Set

pub struct Set<'clipboard> { /* private fields */ }

A builder for an operation that sets a value to the clipboard.

Implementations

impl Set<'_>

pub fn text<'a, T: Into<Cow<'a, str>>>(self, text: T) -> Result<(), Error>

Completes the “set” operation by placing text onto the clipboard. Any valid UTF-8 string is accepted.

pub fn html<'a, T: Into<Cow<'a, str>>>( self, html: T, alt_text: Option<T>, ) -> Result<(), Error>

Completes the “set” operation by placing HTML as well as a plain-text alternative onto the clipboard.

Any valid UTF-8 string is accepted.

pub fn file_list(self, file_list: &[impl AsRef<Path>]) -> Result<(), Error>

Completes the “set” operation by placing a list of file paths onto the clipboard.