arboard
Trait SetExtWindows

pub trait SetExtWindows: Sealed {
    // Required methods
    fn exclude_from_monitoring(self) -> Self;
    fn exclude_from_cloud(self) -> Self;
    fn exclude_from_history(self) -> Self;
}

Windows-specific extensions to the Set builder.

Required Methods

fn exclude_from_monitoring(self) -> Self

Exclude the data which will be set on the clipboard from being processed at all, either in the local clipboard history or getting uploaded to the cloud.

If this is set, it is not recommended to call exclude_from_cloud or exclude_from_history.

fn exclude_from_cloud(self) -> Self

Excludes the data which will be set on the clipboard from being uploaded to the Windows 10/11 cloud clipboard.

fn exclude_from_history(self) -> Self

Excludes the data which will be set on the clipboard from being added to the system’s clipboard history list.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.