Crate arboard

Modules

common 🔒

platform 🔒

Structs

Clear

A builder for an operation that clears the data from the clipboard.

Clipboard

The OS independent struct for accessing the clipboard.

Get

A builder for an operation that gets a value from the clipboard.

ImageData

Stores pixel data of an image.

Set

A builder for an operation that sets a value to the clipboard.

Enums

Error

An error that might happen during a clipboard operation.

LinuxClipboardKind

Clipboard selection

Traits

ClearExtLinux

Linux specific extensions to the Clear builder.

GetExtLinux

Linux-specific extensions to the Get builder.

SetExtLinux

Linux specific extensions to the Set builder.